Patch around compilers that do not support "long long int". (CVS 1656)
FossilOrigin-Name: d98b1502e2947d24ab9f4a5d2e5b9a95dde92faf

+++ src/vdbeInt.h
@@ -385,7 +385,7 @@ int sqlite3VdbeMemCopy(Mem*, const Mem*);
 int sqlite3VdbeMemMove(Mem*, Mem*);
 int sqlite3VdbeMemNulTerminate(Mem*);
 int sqlite3VdbeMemSetStr(Mem*, const char*, int, u8, void(*)(void*));
-void sqlite3VdbeMemSetInt64(Mem*, long long int);
+void sqlite3VdbeMemSetInt64(Mem*, i64);
 void sqlite3VdbeMemSetDouble(Mem*, double);
 void sqlite3VdbeMemSetNull(Mem*);
 int sqlite3VdbeMemMakeWriteable(Mem*);
